requirements.txt
Adafruit-Blinka==4.5.0
adafruit-circuitpython-neopixel==5.1.1
Adafruit-PlatformDetect==2.6.0
Adafruit-PureIO==1.1.4
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.2
Pillow==7.1.1
pyftdi==0.50.1
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
pyusb==1.0.2
six==1.14.0
webcolors==1.11.1
